fix: Set default values with containerManager.
